Same Code, two list display and search
Same code, I added sample data to chemical compounds,

The lists displays based on the button selected, it recycles all the
code with conditional booleans. if Elements is selected then the
periodic table elements are displayed, if compounds are selected then
only compounds are displayed.

I cleaned up a little, there are some things I still want to do with
the code
the xml for displaying can be edited to display element id only when
elements are displayed, and hide it when compounds are called